+++ cedar-policy-cli/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Changed
+
+- The default `--schema-format` is now `human` for all subcommands that take
+  `--schema-format`.
+
 ## 3.1.1
 
 ## 3.1.0